+++ crashserver/server/models/__init__.py
@@ -4,6 +4,7 @@ from .attachments import Attachment
 from .build_metadata import BuildMetadata
 from .minidump import Minidump
 from .project import Project, ProjectType
+from .storage import Storage
 from .symbol import Symbol
 from .symbol_upload import SymbolUploadV2
 from .user import User
@@ -13,6 +14,7 @@ __all__ = [
     "BuildMetadata",
     "Minidump",
     "Project",
+    "Storage",
     "Symbol",
     "SymbolUploadV2",
     "User",