+++ crashserver/server/storage/__init__.py
@@ -1,0 +1,2 @@
+from .storage_factory import register, unregister, get_storage_method, get_storage_methods
+from .storage_target import StorageTarget